tests/wpt/tests/css/css-anchor-position/anchor-position-multicol-001.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<style>
#container {
  position: relative;
  font-family: Ahem;
  font-size: 10px;
  line-height: 1;
  width: 10em;
}
.columns {
  column-width: 100px;
  column-count: 3;
  column-gap: 10px;
  height: 100px;
}
#anchor1 {
  anchor-name: --a1;
  background: blue;
}
.target {
  position: absolute;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div id="container">
    <div>spacer</div>
    <div class="columns">
      <div style="height: 150px"></div>
      <div id="anchor1" style="height: 100px"></div>
    </div>

    <div class="target" style="left: anchor(--a1 left)" data-offset-x=110></div>
    <div class="target" style="right: anchor(--a1 right)" data-offset-x=320></div>
    <div class="target" style="top: anchor(--a1 top)" data-offset-y=10></div>
    <div class="target" style="bottom: anchor(--a1 bottom)" data-offset-y=110></div>
  </div>
</body>
